requirements.txt
# local package
-e .

# external requirements
python-dotenv>=0.5.1
pandas
geopandas
requests
